Move testing code to a tests folder
